Extract SwipeContainerViewController from Manager
